requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cpe-utils==0.1.0
cpeparser==0.0.2
cryptography==41.0.4
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.0.0
ecdsa==0.18.0
fastapi==0.103.1
filelock==3.12.4
h11==0.14.0
idna==3.4
ipwhois==1.2.0
jsonpath-ng==1.6.0
kafka-python==2.0.2
numpy==1.26.2
OTXv2==1.5.12
pandas==2.1.4
paramiko==3.3.1
passlib==1.7.4
pika==1.3.2
pika-stubs==0.1.3
ply==3.11
psycopg2==2.9.7
pyasn1==0.5.0
pycountry==22.3.5
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.4.0
pydantic_core==2.10.0
PyNaCl==1.5.0
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
requests-file==1.5.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tldextract==3.6.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
validators==0.22.0
